Add set generator; convert generic set to be generated (#1)
This introduces a generator for sets of arbitrary Go data types and converts
the top-level generic set to be generated in this way. There are a large number
of build system changes to support the large amount of generated code. The
generated sets are currently intended to have no dependencies on the top-level
module of this package, and have only one external dependency to a well-known
test assert library (github.com/stretchr/testify).

We use multiple Go modules to separate the imports required for the generator
from the imports necessary to use the canned generated set genericset
package.

Special care is taken to continue to support the v1 API of the genericset in
the top-level package.

This doesn't support non-primitive types until #2 is completed.